Relocate all code to new project within solution
