+++ CheckInService/CheckInService.csproj
@@ -17,4 +17,8 @@
     <PackageReference Include="Swashbuckle.AspNetCore" Version="6.4.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\RabbitMQ.Messages\RabbitMQ.Messages.csproj" />
+  </ItemGroup>
+
 </Project>
