+++ RabbitMQ.Messages/Program.cs
@@ -1,1 +1,2 @@
 ﻿// See https://aka.ms/new-console-template for more information
+Console.WriteLine("Hello, World!");
